+++ CHANGELOG.md
@@ -16,6 +16,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## 2019-06-27 
+
+### Changed [Sync-RubrikAnnotation]
+
+* Added -DetailedObject to Get-RubrikVM in order to return the snapshots
+* Added a third annotation to store the date of the latest Rubrik snapshot.
+* Added associated unit tests for Sync-RubrikAnnotation
+
 ## 2019-06-26
 
 ### Added [Tests for Get-RubrikHost]